M4: plugins, artifact backends, fan-out, monitoring endpoint, 0.1.0
Ecosystem work, plus the packaging and lint baseline a public 0.1.0 needs.

Plugins (src/pyattacker/plugins.py)
- Tasks, acquisition algorithms, codecs and stores are discovered through
  `importlib.metadata` entry points, so a third-party package can publish them and
  `use: my_judge` / `algorithm: my_algo` / `store: "s3://bucket/runs.db"` just work.
- Built-ins resolve first (a plugin cannot shadow `echo`), and a plugin that fails to import
  is recorded and skipped instead of breaking the run; `pyattacker plugins` shows both lists.
- examples/plugin_package/ is a real installable plugin (two tasks, an algorithm, a codec)
  used to verify the whole path: install -> discover -> run -> uninstall.

Artifact backends (src/pyattacker/backends.py)
- inline (default) / content-addressed files / null. A `blob_ref` is recorded next to
  digest/size/codec and payloads are hydrated on read, so an artifact stays one object to
  everything upstream and a resumed run reuses spilled checkpoints transparently.
- A schema migration adds the column to stores created before it existed.

Fan-out (pyattacker.fanout)
- Run several tasks on the same input concurrently *inside one task*, returning
  {task_name: value}. Group-level retry is the documented price of not turning pipelines into
  a DAG; first-class Parallel/Gather nodes stay out of scope on purpose.

Monitoring endpoint (src/pyattacker/server.py)
- `pyattacker serve runs/qa.db`: a stdlib-only, read-only HTTP view with JSON endpoints
  (/stats, /events, /pipelines, /resources, /errors) and a small dashboard. It opens a fresh
  read-only connection per request, so it runs happily beside a live run.
- Unauthenticated and loopback-only by default: a debug view, not a service.

Packaging and quality
- Version 0.1.0, MIT LICENSE, `uv build` produces an sdist and a wheel.
- tests/test_packaging.py guards version drift, accidental dependencies, unimportable modules
  and public names that the docs promise but the package does not export.
- ruff is configured in pyproject.toml with a written reason for every ignored rule, and
  `ruff check` is clean across src, tests and examples.
- 242 tests, fully offline; docs/design.md gains the M4 mechanisms and tradeoffs.

+++ src/pyattacker/monitor.py
@@ -9,7 +9,8 @@ from __future__ import annotations
 
 import asyncio
 import time
-from typing import Any, Callable, Mapping
+from collections.abc import Callable, Mapping
+from typing import Any
 
 __all__ = ["render_snapshot", "watch", "read_snapshot"]
 
@@ -20,7 +21,7 @@ _EMPTY = "·"
 def _bar(value: int, total: int, width: int = 20) -> str:
     if total <= 0:
         return _EMPTY * width
-    filled = int(round(width * value / total))
+    filled = round(width * value / total)
     return _BAR * filled + _EMPTY * (width - filled)
